// src/types/input.d.ts
import { InputHTMLAttributes, ReactNode } from "react";

export type InputSize = "sm" | "md" | "lg" | "xl";
export type StyleState = "default" | "invalid";

export interface InputStyleProps {
	size?: InputSize;
	styleState?: StyleState;
	fullWidth?: boolean;
}

export interface BaseInputProps
	extends Omit<InputHTMLAttributes<HTMLInputElement>, "size">,
		InputStyleProps {
	icon?: ReactNode;
	error?: string;
	helper?: string;
	label?: string;
	required?: boolean;
	containerClassName?: string;
	rightElement?: ReactNode;
}